Translator/requirements.txt
streamlit
langchain_core
langchain_community
langchain_text_splitters
langchain_ollama
python-dotenv
langserve[all]